fix(BE): updated all assessments methods and routes

+++ backend/routes/api.php
@@ -73,8 +73,8 @@ Route::prefix("assessment")->group(function () {
     Route::post('/create', [AssessmentController::class, 'createAssessment'])->middleware("isloggedin", "isadmin");
     // Route::get('/{assessmentId}/notify/{employeeId}', [AssessmentController::class, 'notifyEmployeeAssessment']); // do not uncomment this route, some adjusments is currently been made
     Route::get('/{organisationId}', [AssessmentController::class, 'getAssByOrgId']);
-    Route::put('/{assessmentId}', [AssessmentController::class, 'updateAssessment']);
-    Route::delete('/{assessmentId}/delete', [AssessmentController::class, 'deleteAssessment']);
+    Route::put('/{assessmentId}', [AssessmentController::class, 'updateAssessment'])->middleware("isloggedin", "isadmin");
+    Route::delete('/{assessmentId}/delete', [AssessmentController::class, 'deleteAssessment'])->middleware("isloggedin", "isadmin");
 
 });
